Add bash profile
This file defines the settings for a user when running a login
shell. Recall that a shell can be interactive or non-interactive.
A login shell is invoked when a user logins to the terminal either
remotely via SSH or locally, or when bash is started with the
--login option. By contrast, an interactive non-login shell is
invoked when bash is typed into the terminal or when opening a new
terminal tab.

On invocation, Bash reads from

1. ~/.bash_profile
2. ~/.bash_login
3. ~/.profile

in the listed order and executes commands from ~/.bashrc (if that
file exists and is readable). Use the bash_profile file to run
commands that should only execute once, such as customizing
environment variables.